+++ app/src/main/java/comsitejustdoitwhistle/google/sites/monermoney/AddOutcome.java
@@ -18,7 +18,7 @@ import java.util.Calendar;
 public class AddOutcome extends AppCompatActivity {
     private final int TYPE_OUTCOME = 1;
     private Button mCalendarPicker;
-    private int amount;
+    private double amount;
     private Editable comment;
     private int date,month,years;
     private static final int DIALOG_ID = 0;
@@ -76,7 +76,7 @@ public class AddOutcome extends AppCompatActivity {
                 check = true;
                 comment = mNoteTextView.getText();
                 try {
-                    amount = Integer.parseInt(String.valueOf(mAmountIncome.getText()));
+                    amount = Double.parseDouble(String.valueOf(mAmountIncome.getText()));
                 } catch (NumberFormatException e) {
                     check = false;
                     alertDialog.setTitle("ผิดพลาด");
