+++ app/src/main/java/comsitejustdoitwhistle/google/sites/monermoney/Income.java
@@ -1,18 +1,15 @@
 package comsitejustdoitwhistle.google.sites.monermoney;
 
-/**
- * Created by Kenn on 11/18/2016.
- */
 
 public class Income {
-    int money;
+    double money;
     String date, note;
-    public Income(int money, String date, String note) {
+    public Income(double money, String date, String note) {
         this.money = money;
         this.date = date;
         this.note = note;
     }
-    public int getMoney() { return this.money; }
+    public double getMoney() { return this.money; }
     public String getDate() { return this.date; }
     public String getNote() { return this.note; }
 }
